Created generative models object in models subpackage and removed old one in main directory.
